+++ packages/typespec-ts/src/modular/buildCodeModel.ts
@@ -38,7 +38,8 @@ import {
   getProjectedName,
   isNullType,
   getEncode,
-  isTemplateDeclarationOrInstance
+  isTemplateDeclarationOrInstance,
+  UsageFlags
 } from "@typespec/compiler";
 import {
   getAuthentication,
@@ -195,13 +196,18 @@ function isDiscriminator(
   return false;
 }
 
-function handleDiscriminator(context: SdkContext, type: Model) {
+function handleDiscriminator(
+  context: SdkContext,
+  type: Model,
+  usage: UsageFlags
+) {
   const discriminator = getDiscriminator(context.program, type);
   if (discriminator) {
     const discriminatorValues: string[] = [];
     const aliases: string[] = [];
+    const discriminatedSubtypes: Type[] = [];
     for (const childModel of type.derivedModels) {
-      const modelType = getType(context, childModel);
+      const modelType = getType(context, childModel, { usage });
       aliases.push(modelType.name);
       for (const property of modelType.properties) {
         if (property.restApiName === discriminator.propertyName) {
@@ -209,6 +215,7 @@ function handleDiscriminator(context: SdkContext, type: Model) {
           discriminatorValues.push(modelType.discriminatorValue);
         }
       }
+      discriminatedSubtypes.push(modelType);
     }
     const discriminatorInfo = {
       description:
@@ -223,7 +230,8 @@ function handleDiscriminator(context: SdkContext, type: Model) {
       name: discriminator.propertyName,
       isPolymorphic: true,
       isDiscriminator: true,
-      aliases
+      aliases,
+      discriminatedSubtypes
     };
     return discriminatorInfo;
   }
@@ -269,10 +277,11 @@ function getEffectiveSchemaType(program: Program, type: Model | Union): Model {
 function processModelProperties(
   context: SdkContext,
   newValue: any,
-  model: Model
+  model: Model,
+  usage: UsageFlags
 ) {
   // need to do properties after insertion to avoid infinite recursion
-  const discriminatorInfo = handleDiscriminator(context, model);
+  const discriminatorInfo = handleDiscriminator(context, model, usage);
   let hasDiscriminator = false;
   for (const property of model.properties.values()) {
     if (!isSchemaProperty(context.program, property)) {
@@ -284,7 +293,7 @@ function processModelProperties(
     if (newValue.properties === undefined || newValue.properties === null) {
       newValue.properties = [];
     }
-    let newProperty = emitProperty(context, property);
+    let newProperty = emitProperty(context, property, usage);
     if (isDiscriminator(context, model, property.name)) {
       hasDiscriminator = true;
       newProperty = {
@@ -299,10 +308,18 @@ function processModelProperties(
     if (!hasDiscriminator) {
       newValue.properties.push({ ...discriminatorInfo });
     }
-    newValue.alias = `${newValue.name}Parent`;
-    newValue.isPolyBaseModel = true;
-    discriminatorInfo?.aliases.push(`${newValue.alias}`);
-    newValue.aliasType = discriminatorInfo?.aliases.join(" | ");
+    // we don't need to add the discriminator info if it's an anonymous model
+    // because it's impossible to have a anonymous model as the polymorphic base in typespec
+    // the only possibility is the anonymous model is an alias for an union type which has already been taken care of in the combined types.
+    if (newValue.name) {
+      discriminatorInfo?.aliases.push(`${newValue.name}`);
+      newValue.alias = `${newValue.name}`;
+      newValue.name = `${newValue.name}Union`;
+      newValue.aliasType = discriminatorInfo?.aliases.join(" | ");
+      newValue.types = discriminatorInfo?.discriminatedSubtypes;
+      newValue.isPolymorphicBaseModel = true;
+      newValue.discriminator = discriminatorInfo.restApiName;
+    }
   }
 }
 
@@ -321,7 +338,7 @@ function isEmptyAnonymousModel(type: EmitterType): boolean {
 function getType(
   context: SdkContext,
   type: EmitterType,
-  options: { disableEffectiveModel?: boolean } = {}
+  options: { disableEffectiveModel?: boolean; usage?: UsageFlags } = {}
 ): any {
   // don't cache simple type(string, int, etc) since decorators may change the result
   const enableCache = !isSimpleType(context.program, type);
@@ -341,7 +358,7 @@ function getType(
     // do not generate model for empty model, treat it as any
     newValue = { type: "any" };
   } else {
-    newValue = emitType(context, type);
+    newValue = emitType(context, type, options.usage!);
   }
   if (type.kind === "ModelProperty" || type.kind === "Scalar") {
     newValue = applyEncoding(context.program, type, newValue);
@@ -354,13 +371,13 @@ function getType(
     if (type.kind === "Union") {
       for (const t of type.variants.values()) {
         if (t.type.kind === "Model") {
-          processModelProperties(context, newValue, t.type);
+          processModelProperties(context, newValue, t.type, options.usage!);
         }
       }
     }
     if (type.kind === "Model") {
       // need to do properties after insertion to avoid infinite recursion
-      processModelProperties(context, newValue, type);
+      processModelProperties(context, newValue, type, options.usage!);
     }
   } else {
     const key = JSON.stringify(newValue);
@@ -441,7 +458,8 @@ function emitBodyParameter(
     contentTypes = ["application/json"];
   }
   const type = getType(context, getBodyType(context.program, httpOperation)!, {
-    disableEffectiveModel: true
+    disableEffectiveModel: true,
+    usage: UsageFlags.Input
   });
 
   return {
@@ -460,7 +478,9 @@ function emitParameter(
   implementation: string
 ): Parameter {
   const base = emitParamBase(context.program, parameter.param);
-  let type = getType(context, parameter.param.type);
+  let type = getType(context, parameter.param.type, {
+    usage: UsageFlags.Input
+  });
   let clientDefaultValue = undefined;
   if (
     parameter.name.toLowerCase() === "content-type" &&
@@ -559,7 +579,7 @@ function emitResponseHeaders(
   }
   for (const [key, value] of Object.entries(headers)) {
     retval.push({
-      type: getType(context, value.type),
+      type: getType(context, value.type, { usage: UsageFlags.Output }),
       restApiName: key
     });
   }
@@ -588,9 +608,11 @@ function emitResponse(
       candidate.find((e) => e === originType.name)
     ) {
       const modelType = getEffectiveSchemaType(context.program, originType);
-      type = getType(context, modelType);
+      type = getType(context, modelType, { usage: UsageFlags.Output });
     } else {
-      type = getType(context, innerResponse.body.type);
+      type = getType(context, innerResponse.body.type, {
+        usage: UsageFlags.Output
+      });
     }
   }
   const statusCodes: (number | "default")[] = [];
@@ -960,7 +982,8 @@ function isReadOnly(program: Program, type: ModelProperty): boolean {
 
 function emitProperty(
   context: SdkContext,
-  property: ModelProperty
+  property: ModelProperty,
+  usage: UsageFlags
 ): Record<string, any> {
   const newProperty = applyEncoding(context.program, property, property);
   let clientDefaultValue = undefined;
@@ -984,9 +1007,9 @@ function emitProperty(
     getProjectedName(context.program, property, "json") ?? property.name;
 
   if (property.model) {
-    getType(context, property.model);
+    getType(context, property.model, { usage });
   }
-  const type = getType(context, property.type);
+  const type = getType(context, property.type, { usage });
   return {
     clientName: applyCasing(clientName, { casing: CASING }),
     restApiName: jsonName,
@@ -1022,12 +1045,16 @@ function getName(program: Program, type: Model): string {
   }
 }
 
-function emitModel(context: SdkContext, type: Model): Record<string, any> {
+function emitModel(
+  context: SdkContext,
+  type: Model,
+  usage: UsageFlags
+): Record<string, any> {
   // Now we know it's a defined model
   const properties: Record<string, any>[] = [];
   let baseModel = undefined;
   if (type.baseModel) {
-    baseModel = getType(context, type.baseModel);
+    baseModel = getType(context, type.baseModel, { usage });
   }
   const effectiveName = getEffectiveSchemaType(context.program, type).name;
   const overridedModelName =
@@ -1075,14 +1102,15 @@ function emitModel(context: SdkContext, type: Model): Record<string, any> {
     name: modelName,
     description: getDocStr(context.program, type),
     parents: baseModel ? [baseModel] : [],
-    discriminatedSubtypes: {},
+    discriminatedSubtypes: [],
     properties: properties,
     addedOn: getAddedOnVersion(context.program, type),
     snakeCaseName: modelName
       ? applyCasing(modelName, { casing: CASING })
       : modelName,
     base: modelName === "" ? "json" : "dpg",
-    isCoreErrorType: isAzureCoreErrorType(type)
+    isCoreErrorType: isAzureCoreErrorType(type),
+    usage
   };
 }
 
@@ -1334,7 +1362,8 @@ function emitScalar(program: Program, scalar: Scalar): Record<string, any> {
 
 function emitListOrDict(
   context: SdkContext,
-  type: Model
+  type: Model,
+  usage: UsageFlags
 ): Record<string, any> | undefined {
   if (type.indexer !== undefined) {
     if (!isNeverType(type.indexer.key)) {
@@ -1342,12 +1371,12 @@ function emitListOrDict(
       if (name === "string") {
         return {
           type: "dict",
-          elementType: getType(context, type.indexer.value!)
+          elementType: getType(context, type.indexer.value!, { usage })
         };
       } else if (name === "integer") {
         return {
           type: "list",
-          elementType: getType(context, type.indexer.value!)
+          elementType: getType(context, type.indexer.value!, { usage })
         };
       }
     }
@@ -1355,7 +1384,11 @@ function emitListOrDict(
   return undefined;
 }
 
-function mapTypeSpecType(context: SdkContext, type: Type): any {
+function mapTypeSpecType(
+  context: SdkContext,
+  type: Type,
+  usage: UsageFlags
+): any {
   switch (type.kind) {
     case "Number":
       return constantType(type.value, intOrFloat(type.value));
@@ -1364,11 +1397,15 @@ function mapTypeSpecType(context: SdkContext, type: Type): any {
     case "Boolean":
       return constantType(type.value, "boolean");
     case "Model":
-      return emitListOrDict(context, type);
+      return emitListOrDict(context, type, usage);
   }
 }
 
-function emitUnion(context: SdkContext, type: Union): Record<string, any> {
+function emitUnion(
+  context: SdkContext,
+  type: Union,
+  usage: UsageFlags
+): Record<string, any> {
   const sdkType = getSdkUnion(context, type);
   const nonNullOptions = getNonNullOptions(type);
   if (sdkType === undefined) {
@@ -1379,7 +1416,7 @@ function emitUnion(context: SdkContext, type: Union): Record<string, any> {
     const discriminatorPropertyName = getDiscriminator(context.program, type)
       ?.propertyName;
     const variantTypes = sdkType.values.map((x) => {
-      const valueType = getType(context, x.__raw!);
+      const valueType = getType(context, x.__raw!, { usage });
       if (valueType.properties && discriminatorPropertyName) {
         valueType.discriminatorValue = valueType.properties.filter(
           (p: Property) => p.clientName === discriminatorPropertyName
@@ -1395,6 +1432,7 @@ function emitUnion(context: SdkContext, type: Union): Record<string, any> {
       type: "combined",
       types: variantTypes,
       xmlMetadata: {},
+      usage,
       discriminator: discriminatorPropertyName,
       alias:
         unionName === "" || unionName === undefined ? undefined : unionName,
@@ -1413,7 +1451,8 @@ function emitUnion(context: SdkContext, type: Union): Record<string, any> {
       valueType: emitSimpleType(context, sdkType.valueType as SdkBuiltInType),
       values: sdkType.values.map((x) => emitEnumMember(context, x)),
       isFixed: sdkType.isFixed,
-      xmlMetadata: {}
+      xmlMetadata: {},
+      usage
     };
   } else if (
     isStringOrNumberKind(context.program, sdkType.kind) &&
@@ -1432,7 +1471,10 @@ function emitUnion(context: SdkContext, type: Union): Record<string, any> {
       xmlMetadata: {}
     };
   } else {
-    return { ...emitType(context, sdkType.__raw!), nullable: sdkType.nullable };
+    return {
+      ...emitType(context, sdkType.__raw!, usage),
+      nullable: sdkType.nullable
+    };
   }
 }
 
@@ -1491,14 +1533,18 @@ function emitSimpleType(
   };
 }
 
-function emitType(context: SdkContext, type: EmitterType): Record<string, any> {
+function emitType(
+  context: SdkContext,
+  type: EmitterType,
+  usage: UsageFlags
+): Record<string, any> {
   if (type.kind === "Credential") {
     return emitCredential(type.scheme);
   }
   if (type.kind === "CredentialTypeUnion") {
     return emitCredentialUnion(type);
   }
-  const builtinType = mapTypeSpecType(context, type);
+  const builtinType = mapTypeSpecType(context, type, usage);
   if (builtinType !== undefined) {
     // add in description elements for types derived from primitive types (SecureString, etc.)
     const doc = getDoc(context.program, type);
@@ -1512,11 +1558,11 @@ function emitType(context: SdkContext, type: EmitterType): Record<string, any> {
     case "Intrinsic":
       return { type: type.name };
     case "Model":
-      return emitModel(context, type);
+      return emitModel(context, type, usage);
     case "Scalar":
       return emitScalar(context.program, type);
     case "Union":
-      return emitUnion(context, type);
+      return emitUnion(context, type, usage);
     case "UnionVariant":
       return {};
     case "Enum":
@@ -1735,7 +1781,7 @@ function emitCredentialParam(
         };
       }
       return {
-        type: getType(context, type),
+        type: getType(context, type, { usage: UsageFlags.Input }),
         optional: false,
         description: "Credential needed for the client to connect to Azure.",
         clientName: "credential",
@@ -1814,7 +1860,9 @@ function emitClients(
       url: server ? server.url : "",
       apiVersions: [],
       rlcClientName: rlcModels ? getClientName(rlcModels) : client.name,
-      subfolder: ""
+      subfolder: "",
+      rlcHelperDetails:
+        rlcModels && rlcModels.helperDetails ? rlcModels.helperDetails : {}
     };
     const emittedApiVersionParam = getApiVersionParameter(context);
     if (emittedApiVersionParam) {
@@ -1872,7 +1920,7 @@ export function emitCodeModel(
   simpleTypesMap.clear();
   const allModels = getAllModels(dpgContext);
   for (const model of allModels) {
-    getType(dpgContext, model.__raw!);
+    getType(dpgContext, model.__raw!, { usage: model.usage });
   }
 
   for (const namespace of getNamespaces(dpgContext)) {